+++ contracts/test/TickEchidnaTest.sol
@@ -20,7 +20,7 @@ contract TickEchidnaTest {
         // divisibility
         assert((maxTick - minTick) % tickSpacing == 0);
 
-        uint256 numTicks = uint256((maxTick - minTick) / tickSpacing) + 1;
+        uint256 numTicks = uint256(int256((maxTick - minTick) / tickSpacing)) + 1;
         // max liquidity at every tick is less than the cap
         assert(uint256(maxLiquidityPerTick) * numTicks <= type(uint128).max);
     }
